Move field comment generation to base visitor
